layout/style/test/test_pseudo_display_fixup.html

<!doctype html>
<meta charset=utf-8>
<title>Test item blockification of pseudo-elements</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #test {
    display: flex;
  }
  #test::before, #test::after {
    content: "test";
    display: inline-block;
    color: green;
    /*
     * NOTE(emilio): The transition rule is very intentional, to avoid testing
     * the eagerly resolved style.
     */
    transition: color 1s ease;
  }
</style>
<div id="test"></div>
<script>
test(function() {
  document.body.offsetTop;
  let test = document.getElementById("test");
  assert_equals(getComputedStyle(test, "::before").display, "block");
  assert_equals(getComputedStyle(test, "::after").display, "block");
}, "::before and ::after pseudo-elements are blockified");
</script>
